requirements.txt
# Password Manager Dependencies
# CS5371 Cryptography Project

# =============================================================================
# Core Cryptography
# =============================================================================
cryptography>=41.0.0      # AES-GCM encryption
argon2-cffi>=23.1.0       # Argon2id key derivation

# =============================================================================
# Backend (FastAPI)
# =============================================================================
fastapi>=0.109.0          # Web framework
uvicorn[standard]>=0.27.0 # ASGI server
sqlalchemy>=2.0.0         # Database ORM
python-jose[cryptography]>=3.3.0  # JWT tokens
pydantic>=2.0.0           # Data validation

# =============================================================================
# Frontend (PySide6)
# =============================================================================
PySide6>=6.6.0            # Qt for Python GUI

# =============================================================================
# Networking
# =============================================================================
requests>=2.31.0          # HTTP client for API calls

# =============================================================================
# Development (optional)
# =============================================================================
# pytest>=7.4.0           # Testing
# black>=23.0.0           # Code formatting
# mypy>=1.0.0             # Type checking
